+++ cmd/validator/expectation/output.go
@@ -1,4 +1,4 @@
-// Copyright © 2021 Weald Technology Trading.
+// Copyright © 2021, 2023 Weald Technology Trading.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -15,24 +15,42 @@ package validatorexpectation
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
 	"strings"
 
 	"github.com/hako/durafmt"
 )
 
-func (c *command) output(_ context.Context) (string, error) {
+func (c *command) output(ctx context.Context) (string, error) {
 	if c.quiet {
 		return "", nil
 	}
 
+	if c.json {
+		return c.outputJSON(ctx)
+	}
+	return c.outputTxt(ctx)
+}
+
+func (c *command) outputJSON(_ context.Context) (string, error) {
+	data, err := json.Marshal(c.res)
+	if err != nil {
+		return "", err
+	}
+
+	return fmt.Sprintf("%s\n", string(data)), nil
+}
+
+func (c *command) outputTxt(_ context.Context) (string, error) {
 	builder := strings.Builder{}
 
 	builder.WriteString("Expected time between block proposals: ")
-	builder.WriteString(durafmt.Parse(c.timeBetweenProposals).LimitFirstN(2).String())
+	builder.WriteString(durafmt.Parse(c.res.timeBetweenProposals).LimitFirstN(2).String())
 	builder.WriteString("\n")
 
 	builder.WriteString("Expected time between sync committees: ")
-	builder.WriteString(durafmt.Parse(c.timeBetweenSyncCommittees).LimitFirstN(2).String())
+	builder.WriteString(durafmt.Parse(c.res.timeBetweenSyncCommittees).LimitFirstN(2).String())
 	builder.WriteString("\n")
 
 	return builder.String(), nil