+++ CHANGELOG.md
@@ -3,6 +3,7 @@ dev:
   - add "--generate-keystore" option for "account derive"
   - update "validator exit" command to be able to generate multiple exits
   - support for 12-word and 18-word mnemonics with single-word (no whitespace) passphrases
+  - add JSON output for "validator expectation"
 
 1.30.0:
   - add "chain spec" command